requirements.txt
Django==4.2.6
djangorestframework==3.14.0
djongo==1.3.6
djangotoolbox==1.6.2
djangorestframework_simplejwt==5.3.0
